requirements.txt
aioredis==2.0.1

python-telegram-bot==20.2